llm-agent/requirements.txt
# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langgraph>=0.0.40
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
typing-extensions>=4.5.0

# Dependencies from agentic-pipeline (required for tools)
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
pyyaml>=6.0
colorlog>=6.7.0
jsonschema>=4.17.0
